requirements.txt
groq>=0.9.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.7.0
rich>=13.0.0
typer>=0.12.0
httpx>=0.27.0
